feat: NativeScript 4.0 integration
Angular application is started with `application.run()` which means the the root view of the app will be the root view of the `AppComponent`.
Previous behavior: Anagular application was started with `application.start()` - this aways creates a root `Frame`. The platfrom bootstrap was always creating the inital `Page` and loading the `AppComponent` inside.

The `<page-router-outlet>` will create a Frame as a native element and will create a page for each component activated in it (including the inital page).